+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * #84 headless? bug fixed
   * #79 readme updated, file upload section
   * #16 postmortem filepath
+  * #69 dump console logs in PM
 
 ## 0.2.2
   * frames and iframes support